First real build of assignment 1
The very basics of what we need for our assignment. A player character
that can move with tank controls, 2 enemy agents, and one wall. We need
one more wall, clean up the placements and sprites themselves, and get
the player sensors working, but all that can be done fairly easily I
think. The sprites will need to be smaller and what we have now is just
placeholders